adapter: remove CopyRows command
We already had all the primitives we needed to process the copy protocol
without going through the coordinator. Refactor and move things around
to make those available to the adapter.
