+++ src/adapter/src/coord/sequencer.rs
@@ -14,26 +14,27 @@
 
 use inner::return_if_err;
 use mz_controller::clusters::ClusterId;
-use mz_expr::OptimizedMirRelationExpr;
+use mz_expr::{MirRelationExpr, OptimizedMirRelationExpr, RowSetFinishing};
 use mz_ore::tracing::OpenTelemetryContext;
 use mz_repr::explain::ExplainFormat;
-use mz_repr::{GlobalId, Timestamp};
-use mz_sql::catalog::CatalogCluster;
+use mz_repr::{Diff, GlobalId, Timestamp};
+use mz_sql::catalog::{CatalogCluster, CatalogError};
 use mz_sql::names::ResolvedIds;
 use mz_sql::plan::{
-    AbortTransactionPlan, CommitTransactionPlan, CopyRowsPlan, CreateRolePlan, CreateSourcePlans,
-    FetchPlan, Params, Plan, PlanKind, RaisePlan, RotateKeysPlan,
+    self, AbortTransactionPlan, CommitTransactionPlan, CreateRolePlan, CreateSourcePlans,
+    FetchPlan, MutationKind, Params, Plan, PlanKind, RaisePlan, RotateKeysPlan,
 };
 use mz_sql_parser::ast::{Raw, Statement};
 use tokio::sync::oneshot;
 use tracing::{event, Level};
 
+use crate::catalog::Catalog;
 use crate::command::{Command, ExecuteResponse, Response};
 use crate::coord::id_bundle::CollectionIdBundle;
 use crate::coord::{introspection, Coordinator, Message};
 use crate::error::AdapterError;
 use crate::notice::AdapterNotice;
-use crate::session::{EndTransactionAction, Session, TransactionStatus};
+use crate::session::{EndTransactionAction, Session, TransactionOps, TransactionStatus, WriteOp};
 use crate::util::ClientTransmitter;
 use crate::{rbac, ExecuteContext, ExecuteResponseKind};
 
@@ -298,9 +299,6 @@ impl Coordinator {
                     }),
                     session,
                 );
-            }
-            Plan::CopyRows(CopyRowsPlan { id, columns, rows }) => {
-                self.sequence_copy_rows(ctx, id, columns, rows);
             }
             Plan::Explain(plan) => {
                 self.sequence_explain(ctx, plan, target_cluster).await;
@@ -646,5 +644,116 @@ impl Coordinator {
                 session.add_notice(AdapterNotice::RbacUserDisabled);
             }
         }
+    }
+
+    pub(crate) fn insert_constant(
+        catalog: &Catalog,
+        session: &mut Session,
+        id: GlobalId,
+        constants: MirRelationExpr,
+    ) -> Result<ExecuteResponse, AdapterError> {
+        // Insert can be queued, so we need to re-verify the id exists.
+        let desc = match catalog.try_get_entry(&id) {
+            Some(table) => {
+                table.desc(&catalog.resolve_full_name(table.name(), Some(session.conn_id())))?
+            }
+            None => {
+                return Err(AdapterError::SqlCatalog(CatalogError::UnknownItem(
+                    id.to_string(),
+                )))
+            }
+        };
+
+        match constants.as_const() {
+            Some((rows, ..)) => {
+                let rows = rows.clone()?;
+                for (row, _) in &rows {
+                    for (i, datum) in row.iter().enumerate() {
+                        desc.constraints_met(i, &datum)?;
+                    }
+                }
+                let diffs_plan = plan::SendDiffsPlan {
+                    id,
+                    updates: rows,
+                    kind: MutationKind::Insert,
+                    returning: Vec::new(),
+                    max_result_size: catalog.system_config().max_result_size(),
+                };
+                Self::send_diffs(session, diffs_plan)
+            }
+            None => panic!(
+                "tried using sequence_insert_constant on non-constant MirRelationExpr {:?}",
+                constants
+            ),
+        }
+    }
+
+    #[tracing::instrument(level = "debug", skip_all)]
+    pub(crate) fn send_diffs(
+        session: &mut Session,
+        mut plan: plan::SendDiffsPlan,
+    ) -> Result<ExecuteResponse, AdapterError> {
+        let affected_rows = {
+            let mut affected_rows = Diff::from(0);
+            let mut all_positive_diffs = true;
+            // If all diffs are positive, the number of affected rows is just the
+            // sum of all unconsolidated diffs.
+            for (_, diff) in plan.updates.iter() {
+                if *diff < 0 {
+                    all_positive_diffs = false;
+                    break;
+                }
+
+                affected_rows += diff;
+            }
+
+            if !all_positive_diffs {
+                // Consolidate rows. This is useful e.g. for an UPDATE where the row
+                // doesn't change, and we need to reflect that in the number of
+                // affected rows.
+                differential_dataflow::consolidation::consolidate(&mut plan.updates);
+
+                affected_rows = 0;
+                // With retractions, the number of affected rows is not the number
+                // of rows we see, but the sum of the absolute value of their diffs,
+                // e.g. if one row is retracted and another is added, the total
+                // number of rows affected is 2.
+                for (_, diff) in plan.updates.iter() {
+                    affected_rows += diff.abs();
+                }
+            }
+
+            usize::try_from(affected_rows).expect("positive isize must fit")
+        };
+        event!(
+            Level::TRACE,
+            affected_rows,
+            id = format!("{:?}", plan.id),
+            kind = format!("{:?}", plan.kind),
+            updates = plan.updates.len(),
+            returning = plan.returning.len(),
+        );
+
+        session.add_transaction_ops(TransactionOps::Writes(vec![WriteOp {
+            id: plan.id,
+            rows: plan.updates,
+        }]))?;
+        if !plan.returning.is_empty() {
+            let finishing = RowSetFinishing {
+                order_by: Vec::new(),
+                limit: None,
+                offset: 0,
+                project: (0..plan.returning[0].0.iter().count()).collect(),
+            };
+            return match finishing.finish(plan.returning, plan.max_result_size) {
+                Ok(rows) => Ok(Self::send_immediate_rows(rows)),
+                Err(e) => Err(AdapterError::ResultSize(e)),
+            };
+        }
+        Ok(match plan.kind {
+            MutationKind::Delete => ExecuteResponse::Deleted(affected_rows),
+            MutationKind::Insert => ExecuteResponse::Inserted(affected_rows),
+            MutationKind::Update => ExecuteResponse::Updated(affected_rows / 2),
+        })
     }
 }
